veritas_ai_backend_fastapi/requirements.txt
fastapi
uvicorn[standard]
google-genai
pydantic
python-dotenv
pydantic-settings
